:recycle: Move text serialization code to wasm module
